GUI V2
This version adds many functionalities that are described in the done section of the GUI project. The frame and panel setup is a little messed up and needs urgent work. To make it realistically useable, you must run the program the manually drag the corner of the GUI window to increase its size.